+++ internal/protobuf/buffers.go
@@ -17,24 +17,22 @@ type MemBuffer struct {
 	pool *sync.Pool
 }
 
-// Finalize seals underlying message with given bounds and increments ref
-// counter for x. Finalize is intended to be called before.
-func (x *MemBuffer) Finalize(from, to int) {
+// SetBounds seals underlying message with given bounds.
+func (x *MemBuffer) SetBounds(from, to int) {
 	x.from = from
 	x.to = to
-	x.Ref()
 }
 
-// Len returns final length of the underlying message. Len returns undefined
-// value before [MemBuffer.Finalize].
+// ReadOnlyData returns final length of the underlying message. ReadOnlyData returns undefined
+// value before [MemBuffer.SetBounds].
 //
-// Len implements [mem.Buffer].
+// ReadOnlyData implements [mem.Buffer].
 func (x *MemBuffer) ReadOnlyData() []byte {
 	return x.SliceBuffer[x.from:x.to]
 }
 
 // Len returns final length of the underlying message. Len returns undefined
-// value before [MemBuffer.Finalize].
+// value before [MemBuffer.SetBounds].
 //
 // Len implements [mem.Buffer].
 func (x *MemBuffer) Len() int {
@@ -90,7 +88,7 @@ func NewBufferPool(ln int) *MemBufferPool {
 // Get selects MemBuffer from x and returns it.
 //
 // Instant [MemBuffer.Free] call returns the buffer to x. Use [MemBuffer.Ref] /
-// [MemBuffer.Finalize] to prevent release.
+// to prevent release.
 func (x *MemBufferPool) Get() *MemBuffer {
 	item := x.syncPool.Get().(*MemBuffer)
 	item.pool = x.syncPool