+++ pkg/local_object_storage/blobstor/fstree/head.go
@@ -36,8 +36,28 @@ func (t *FSTree) Head(addr oid.Address) (*object.Object, error) {
 //
 // Passed buf must have 2*[object.MaxHeaderLen] bytes len at least.
 func (t *FSTree) ReadHeader(addr oid.Address, buf []byte) (int, error) {
+	n, stream, err := t.ReadObject(addr, buf)
+	if err != nil {
+		return 0, err
+	}
+
+	stream.Close()
+
+	return n, nil
+}
+
+// ReadObject reads first bytes of the referenced object's binary containing its
+// full header from t into buf. Returns number of bytes read and stream of
+// remaining bytes. The stream must be finally closed by the caller.
+//
+// Read part may include payload prefix.
+//
+// If object is missing, ReadObject returns [apistatus.ErrObjectNotFound].
+//
+// Passed buf must have 2*[object.MaxHeaderLen] bytes len at least.
+func (t *FSTree) ReadObject(addr oid.Address, buf []byte) (int, io.ReadCloser, error) {
 	if len(buf) < 2*object.MaxHeaderLen {
-		return 0, fmt.Errorf("too short buffer %d bytes", len(buf))
+		return 0, nil, fmt.Errorf("too short buffer %d bytes", len(buf))
 	}
 
 	p := t.treePath(addr)
@@ -45,27 +65,27 @@ func (t *FSTree) ReadHeader(addr oid.Address, buf []byte) (int, error) {
 	f, err := os.Open(p)
 	if err != nil {
 		if errors.Is(err, fs.ErrNotExist) {
-			return 0, logicerr.Wrap(apistatus.ErrObjectNotFound)
+			return 0, nil, logicerr.Wrap(apistatus.ErrObjectNotFound)
 		}
-		return 0, fmt.Errorf("read file %q: %w", p, err)
+		return 0, nil, fmt.Errorf("read file %q: %w", p, err)
 	}
 
 	initial, stream, err := t.readHeader(addr.Object(), f, buf)
 	if err != nil {
 		stream.Close()
-		return 0, err
+		return 0, nil, err
 	}
 
 	initial, stream, err = t.preprocessStreamHead(stream, initial)
-	if stream != nil {
-		stream.Close()
-	}
-
 	if err != nil {
-		return 0, err
+		return 0, nil, err
 	}
 
-	return copy(buf, initial), nil
+	if stream == nil {
+		stream = nopReadCloser{}
+	}
+
+	return copy(buf, initial), stream, nil
 }
 
 // getObjectStream reads an object from the storage by address as a stream.