+++ pkg/local_object_storage/blobstor/fstree/util.go
@@ -13,3 +13,9 @@ func (x readerTwoClosers) Close() error {
 	x.ReadCloser.Close()
 	return x.baseCloser.Close()
 }
+
+type nopReadCloser struct{}
+
+func (nopReadCloser) Read([]byte) (int, error) { return 0, io.EOF }
+
+func (nopReadCloser) Close() error { return nil }