+++ pkg/local_object_storage/shard/get_test.go
@@ -2,12 +2,16 @@ package shard_test
 
 import (
 	"bytes"
+	"encoding/binary"
 	"errors"
 	"io"
 	"testing"
+	"testing/iotest"
 	"time"
 
+	iprotobuf "github.com/nspcc-dev/neofs-node/internal/protobuf"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/shard"
+	apistatus "github.com/nspcc-dev/neofs-sdk-go/client/status"
 	cidtest "github.com/nspcc-dev/neofs-sdk-go/container/id/test"
 	"github.com/nspcc-dev/neofs-sdk-go/object"
 	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
@@ -145,4 +149,74 @@ func testGetStream(t testing.TB, sh *shard.Shard, addr oid.Address, obj *object.
 // require.Equal fails on comparing <nil> and []{} lists.
 func binaryEqual(a, b *object.Object) bool {
 	return bytes.Equal(a.Marshal(), b.Marshal())
+}
+
+func TestShard_ReadObject(t *testing.T) {
+	t.Run("without write cache", func(t *testing.T) {
+		testShardReadObject(t, false)
+	})
+
+	t.Run("with write cache", func(t *testing.T) {
+		testShardReadObject(t, true)
+	})
+}
+
+func testShardReadObject(t *testing.T, hasWriteCache bool) {
+	sh := newShard(t, hasWriteCache)
+	defer releaseShard(sh, t)
+
+	buf := make([]byte, 2*object.MaxHeaderLen)
+
+	t.Run("regular object", func(t *testing.T) {
+		obj := generateObject()
+		addAttribute(obj, "foo", "bar")
+
+		err := sh.Put(obj, nil)
+		require.NoError(t, err)
+
+		n, stream, err := sh.ReadObject(obj.Address(), false, buf)
+		require.NoError(t, err)
+		assertReadObjectOK(t, *obj, buf, n, stream)
+	})
+
+	t.Run("virtual object", func(t *testing.T) {
+		cnr := cidtest.ID()
+
+		parent := generateObjectWithCID(cnr)
+		addAttribute(parent, "foo", "bar")
+
+		child := generateObjectWithCID(cnr)
+		child.SetParent(parent)
+
+		err := sh.Put(child, nil)
+		require.NoError(t, err)
+
+		_, _, err = sh.ReadObject(parent.Address(), false, buf)
+		var siErr *object.SplitInfoError
+		require.ErrorAs(t, err, &siErr)
+		si := siErr.SplitInfo()
+		require.NotNil(t, si)
+		require.Equal(t, child.GetID(), si.GetLastPart())
+
+		_, _, err = sh.ReadObject(parent.Address(), true, buf)
+		require.ErrorIs(t, err, apistatus.ErrObjectNotFound)
+	})
+}
+
+func assertReadObjectOK(t *testing.T, obj object.Object, buf []byte, n int, stream io.ReadCloser) {
+	if len(obj.Payload()) > 0 {
+		_, tail, ok := bytes.Cut(buf[:n], obj.CutPayload().Marshal())
+		require.True(t, ok)
+
+		prefix := make([]byte, 1+binary.MaxVarintLen64)
+		prefix[0] = iprotobuf.TagBytes4 // payload field tag
+		prefix = prefix[:1+binary.PutUvarint(prefix[1:], uint64(len(obj.Payload())))]
+
+		tail, ok = bytes.CutPrefix(tail, prefix)
+		require.True(t, ok)
+		require.True(t, bytes.HasPrefix(obj.Payload(), tail))
+	}
+
+	require.NoError(t, iotest.TestReader(io.MultiReader(bytes.NewReader(buf[:n]), stream), obj.Marshal()))
+	require.NoError(t, stream.Close())
 }