+++ pkg/local_object_storage/writecache/writecache.go
@@ -30,6 +30,7 @@ type Cache interface {
 	GetRangeStream(addr oid.Address, off uint64, ln uint64) (io.ReadCloser, error)
 	Head(oid.Address) (*object.Object, error)
 	ReadHeader(oid.Address, []byte) (int, error)
+	ReadObject(oid.Address, []byte) (int, io.ReadCloser, error)
 	// Delete removes object referenced by the given oid.Address from the
 	// Cache. Returns any error encountered that prevented the object to be
 	// removed.